fix: changed replace to use int64

+++ pandas/_libs/tslibs/timedeltas.pxd
@@ -10,8 +10,8 @@ cpdef int64_t delta_to_nanoseconds(
     delta, NPY_DATETIMEUNIT reso=*, bint round_ok=*
 ) except? -1
 cdef convert_to_timedelta64(object ts, str unit)
-cdef create_timedelta_from_parts(int days=*, int hours=*, int minutes=*, int seconds=*,
-    int milliseconds=*, int microseconds=*, int nanoseconds=*)
+cdef create_timedelta_from_parts(int64_t days=*, int64_t hours=*, int64_t minutes=*, int64_t seconds=*,
+    int64_t milliseconds=*, int64_t microseconds=*, int64_t nanoseconds=*)
 cdef bint is_any_td_scalar(object obj)
